+++ src/patterns.ts
@@ -57,6 +57,12 @@ export const valueTypeText = new RegExp(
 export const valueTypeTime = /[0-9]{6}Z?/
 export const valueTypeUtcOffset = /[+-]([0-9]{2}){2,3}/
 
+// Content type as defined by RFC 4288 4.2
+export const regName = /[a-zA-Z0-9!#$&.+^_-]{1,127}/
+export const contentType = new RegExp(
+    '(' + regName.source + ')/(' + regName.source + ')'
+)
+
 /**
  * Check if a string matches a pattern for the whole string.
  * @param pattern The RegExp pattern to match against.
@@ -66,4 +72,62 @@ export const valueTypeUtcOffset = /[+-]([0-9]{2}){2,3}/
 export function matchesWholeString(pattern: RegExp, text: string): boolean {
     const match = text.match(pattern)
     return match !== null && match[0] === text
+}
+
+/**
+ * Get the ordinal (character code) of a character.
+ * @param char The character to get the ordinal value of.
+ * @returns The ordinal as a number.
+ * @throws If the input string is empty.
+ */
+export function ord(char: string): number {
+    if (char.length === 0)
+        throw new Error('Expected a character, got empty string')
+
+    return char.charCodeAt(0)
+}
+
+/**
+ * Check if a character is allowed in a property or parameter name.
+ * @param char The character to check.
+ * @returns Whether or not the character is allowed.
+ */
+export function isNameChar(char: string): boolean {
+    return (
+        (ord(char) >= ord('A') && ord(char) <= ord('Z')) ||
+        (ord(char) >= ord('a') && ord(char) <= ord('z')) ||
+        (ord(char) >= ord('0') && ord(char) <= ord('9')) ||
+        char === '-'
+    )
+}
+
+/**
+ * Check if a character is allowwed in a property value.
+ * @param char The character to check.
+ * @returns Whether or not the character is allowed.
+ */
+export function isPropertyValueChar(char: string): boolean {
+    return ord(char) === 9 || ord(char) >= 32
+}
+
+/**
+ * Check if a character is allowwed in a parameter value.
+ * @param char The character to check.
+ * @param quoted Whether or not the parameter value is quoted, this allows ';', ':' and ','.
+ * @returns Whether or not the character is allowed.
+ */
+export function isParameterValueChar(
+    char: string,
+    quoted: boolean = false
+): boolean {
+    return (
+        // Allow HTAB
+        ord(char) === 9 ||
+        // Do not allow DQUOTE
+        (char !== '"' &&
+            // Allow non-control characters
+            ord(char) >= 32 &&
+            // These characters are only allowed if quoted
+            (quoted || !(char === ';' || char === ':' || char === ',')))
+    )
 }